# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/driver-locations": {
      "post": {
        "operationId": "DriverLocationsController_updateLocation",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateLocationDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiSuccessEnvelope"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "nullable": true
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "The request failed validation. `errors` lists the offending fields.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailureEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "No access token, or the token is invalid or expired.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailureEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated, but not allowed to act on this resource.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailureEnvelope"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailureEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error. The body carries a `requestId` to quote when reporting it; no internal detail is exposed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailureEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "bearer": []
          }
        ],
        "summary": "Publish the calling driver's current GPS position",
        "tags": [
          "Driver Locations"
        ]
      }
    },
    "/api/v1/driver-locations/driver/{driverId}": {
      "get": {
        "operationId": "DriverLocationsController_getDriverLocation",
        "parameters": [
          {
            "name": "driverId",
            "required": true,
            "in": "path",
            "description": "Driver UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Driver location fetched successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiSuccessEnvelope"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "nullable": true
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "The request failed validation. `errors` lists the offending fields.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailureEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "No access token, or the token is invalid or expired.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailureEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated, but not allowed to act on this resource.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailureEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "No such resource.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailureEnvelope"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailureEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error. The body carries a `requestId` to quote when reporting it; no internal detail is exposed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailureEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get location for a specific driver",
        "tags": [
          "Driver Locations"
        ]
      }
    },
    "/api/v1/driver-locations/nearby": {
      "get": {
        "operationId": "DriverLocationsController_findNearbyDrivers",
        "parameters": [
          {
            "name": "lat",
            "required": true,
            "in": "query",
            "description": "Latitude of search point",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "lng",
            "required": true,
            "in": "query",
            "description": "Longitude of search point",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "radius",
            "required": false,
            "in": "query",
            "description": "Search radius in kilometres. Defaults to 10 and is CAPPED at 50; a larger value is silently clamped, and an unparseable one falls back to the default.",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Nearby drivers fetched successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiSuccessEnvelope"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "nullable": true
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "The request failed validation. `errors` lists the offending fields.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailureEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "No access token, or the token is invalid or expired.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailureEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated, but not allowed to act on this resource.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailureEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "No such resource.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailureEnvelope"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailureEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error. The body carries a `requestId` to quote when reporting it; no internal detail is exposed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailureEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Find nearby drivers by coordinates and radius",
        "tags": [
          "Driver Locations"
        ]
      }
    },
...
  },
  "info": {
    "title": "Ark Rides API",
    "description": "Ride-hailing backend for Ark Rides: rider and driver identity (Privy + email), ride lifecycle, fare ledger, driver wallet, emergency SOS and off-app booking channels.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Enter JWT access token"
      }
    },
    "schemas": {
...
      "UpdateLocationDto": {
        "type": "object",
        "properties": {
          "latitude": {
            "type": "number",
            "example": 6.5244,
            "minimum": -90,
            "maximum": 90
          },
          "longitude": {
            "type": "number",
            "example": 3.3792,
            "minimum": -180,
            "maximum": 180
          }
        },
        "required": [
          "latitude",
          "longitude"
        ]
      },
...
      "ApiFieldError": {
        "type": "object",
        "properties": {
          "field": {
            "type": "string",
            "example": "pickup.lat",
            "description": "Dotted path to the offending property."
          },
          "messages": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "example": [
              "lat must be between -90 and 90"
            ]
          }
        },
        "required": [
          "field",
          "messages"
        ]
      },
      "ApiFailureEnvelope": {
        "type": "object",
        "description": "Returned for every 4xx and 5xx. `message` is ALWAYS a single string; field-level detail goes in `errors`.",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "statusCode": {
            "type": "integer",
            "example": 403
          },
          "message": {
            "type": "string",
            "example": "You can only view your own ride history."
          },
          "code": {
            "type": "string",
            "description": "Machine-readable discriminator; branch on this, not on the message.",
            "example": "FORBIDDEN",
            "enum": [
              "VALIDATION_FAILED",
              "UNAUTHENTICATED",
              "FORBIDDEN",
              "NOT_FOUND",
              "CONFLICT",
              "DUPLICATE_VALUE",
              "INVALID_REFERENCE",
              "MALFORMED_IDENTIFIER",
              "RATE_LIMITED",
              "DATABASE_ERROR",
              "INTERNAL_ERROR"
            ]
          },
          "errors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ApiFieldError"
            },
            "description": "Present only for validation failures."
          },
          "path": {
            "type": "string",
            "example": "/api/v1/rides/abc"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          },
          "requestId": {
            "type": "string",
            "description": "Present only on 5xx. Quote it when reporting a fault."
          }
        },
        "required": [
          "success",
          "statusCode",
          "message",
          "code",
          "path",
          "timestamp"
        ]
      },
      "ApiSuccessEnvelope": {
        "type": "object",
        "description": "Returned for every 2xx that carries a body.",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "statusCode": {
            "type": "integer",
            "example": 200
          },
          "message": {
            "type": "string",
            "example": "Request successful"
          },
          "data": {
            "description": "The payload. Its shape is per-endpoint.",
            "nullable": true
          },
          "meta": {
            "type": "object",
            "description": "Present on list endpoints.",
            "properties": {
              "page": {
                "type": "integer",
                "example": 1
              },
              "limit": {
                "type": "integer",
                "example": 20
              },
              "total": {
                "type": "integer",
                "example": 41
              },
              "totalPages": {
                "type": "integer",
                "example": 3
              }
            }
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "success",
          "statusCode",
          "message",
          "data",
          "timestamp"
        ]
      }
    }
  }
}